fix: 统一录像清理与网络归档弹窗风格
两个弹窗改为与 ConfirmDialog 一致的无边框圆角卡片，复用应用主题组件；新增共享 InputTextBoxStyle 输入框主题（圆角、主题色、聚焦高亮）；设置页录像清理卡片改用 SettingRowStyle 标准行布局，标题与按钮同栏；补齐新增 XAML 文案的中英文资源。
