+++ ExpressPackingMonitoring/ExpressPackingMonitoring.csproj
@@ -195,14 +195,14 @@
     <Compile Include="UI\MobileAppUpdatePrompt.cs" />
     <Compile Include="UI\MobileAppUpdatePromptWindow.xaml.cs" />
     <Compile Include="UI\MobileConnectionWindow.xaml.cs" />
-    <Compile Include="UI\ManualCleanupDialog.cs" />
+    <Compile Include="UI\ManualCleanupDialog.xaml.cs" />
     <Compile Include="UI\NotificationVisuals.cs" />
     <Compile Include="UI\PlaybackWindow.xaml.cs" />
     <Compile Include="UI\VideoImportDialog.xaml.cs" />
     <Compile Include="UI\SettingsWindow.xaml.cs" />
     <Compile Include="UI\SettingsContext.cs" />
     <Compile Include="UI\StatisticsWindow.xaml.cs" />
-    <Compile Include="UI\StoragePathSelectionDialog.cs" />
+    <Compile Include="UI\StoragePathSelectionDialog.xaml.cs" />
     <Compile Include="UI\UpdateAvailableDialog.xaml.cs" />
     <Compile Include="Workstations\PrintToolInstallGuide.cs" />
     <Compile Include="Workstations\WorkstationInstanceCoordinator.cs" />
