+++ ExpressPackingMonitoring/Resources/Strings.zh-Hans.resx
@@ -910,4 +910,7 @@
   <data name="清理只影响电脑本地录像，NAS 中已备份的文件不会受到影响" xml:space="preserve"><value>清理只影响电脑本地录像，NAS 中已备份的文件不会受到影响</value></data>
   <data name="按时间清理…" xml:space="preserve"><value>按时间清理…</value></data>
   <data name="按空间释放…" xml:space="preserve"><value>按空间释放…</value></data>
+  <data name="开始清理" xml:space="preserve"><value>开始清理</value></data>
+  <data name="选择录像保存文件夹" xml:space="preserve"><value>选择录像保存文件夹</value></data>
+  <data name="本地文件夹或 UNC 网络路径" xml:space="preserve"><value>本地文件夹或 UNC 网络路径</value></data>
 </root>
