+++ ExpressPackingMonitoring/UI/ManualCleanupDialog.xaml
@@ -1,0 +1,80 @@
+<Window x:Class="ExpressPackingMonitoring.UI.ManualCleanupDialog"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
+        Title="录像清理"
+        SizeToContent="Height"
+        Width="460"
+        MaxHeight="620"
+        WindowStyle="None"
+        AllowsTransparency="True"
+        ResizeMode="NoResize"
+        ShowInTaskbar="False"
+        WindowStartupLocation="CenterOwner"
+        Background="{StaticResource TransparentBrush}"
+        Foreground="{DynamicResource TextPrimary}">
+    <Border Background="{DynamicResource PanelBackground}"
+            BorderBrush="{DynamicResource BorderDefault}"
+            BorderThickness="1"
+            CornerRadius="14"
+            Padding="24">
+        <StackPanel>
+            <TextBlock x:Name="DialogTitleText"
+                       FontSize="18"
+                       FontWeight="Bold"
+                       Foreground="{DynamicResource TextPrimary}"/>
+            <TextBlock Text="清理只影响电脑本地录像，NAS 中已备份的文件不会受到影响"
+                       FontSize="12"
+                       TextWrapping="Wrap"
+                       Foreground="{DynamicResource TextSecondary}"
+                       Margin="0,8,0,16"/>
+            <StackPanel Orientation="Horizontal">
+                <ComboBox x:Name="PresetCombo"
+                          Width="220"
+                          VerticalAlignment="Center"
+                          Cursor="Hand"/>
+                <xctk:IntegerUpDown x:Name="DaysInput"
+                                    Width="110"
+                                    Minimum="1"
+                                    Maximum="3650"
+                                    Value="30"
+                                    Margin="8,0,0,0"
+                                    VerticalAlignment="Center"
+                                    Visibility="Collapsed"/>
+                <xctk:DoubleUpDown x:Name="SpaceInput"
+                                   Width="110"
+                                   Minimum="1"
+                                   Maximum="100000"
+                                   Value="10"
+                                   FormatString="F0"
+                                   Margin="8,0,0,0"
+                                   VerticalAlignment="Center"
+                                   Visibility="Collapsed"/>
+                <TextBlock x:Name="UnitText"
+                           Margin="4,0,0,0"
+                           VerticalAlignment="Center"
+                           Foreground="{DynamicResource TextSecondary}"/>
+            </StackPanel>
+            <TextBlock x:Name="ModeDescriptionText"
+                       FontSize="12"
+                       TextWrapping="Wrap"
+                       Foreground="{DynamicResource TextSecondary}"
+                       Margin="0,10,0,0"/>
+            <StackPanel Orientation="Horizontal"
+                        HorizontalAlignment="Right"
+                        Margin="0,22,0,0">
+                <Button x:Name="OkButton"
+                        MinWidth="92"
+                        Content="开始清理"
+                        Style="{StaticResource DangerButtonStyle}"
+                        Margin="0,0,10,0"
+                        IsDefault="True"
+                        Click="OkButton_Click"/>
+                <Button MinWidth="92"
+                        Content="取消"
+                        Style="{StaticResource SecondaryButtonStyle}"
+                        IsCancel="True"/>
+            </StackPanel>
+        </StackPanel>
+    </Border>
+</Window>
