+++ ExpressPackingMonitoring/UI/StoragePathSelectionDialog.xaml
@@ -1,0 +1,63 @@
+<Window x:Class="ExpressPackingMonitoring.UI.StoragePathSelectionDialog"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="选择录像保存文件夹"
+        SizeToContent="Height"
+        Width="580"
+        MaxHeight="620"
+        WindowStyle="None"
+        AllowsTransparency="True"
+        ResizeMode="NoResize"
+        ShowInTaskbar="False"
+        WindowStartupLocation="CenterOwner"
+        Background="{StaticResource TransparentBrush}"
+        Foreground="{DynamicResource TextPrimary}">
+    <Border Background="{DynamicResource PanelBackground}"
+            BorderBrush="{DynamicResource BorderDefault}"
+            BorderThickness="1"
+            CornerRadius="14"
+            Padding="24">
+        <StackPanel>
+            <TextBlock x:Name="DialogTitleText"
+                       FontSize="18"
+                       FontWeight="Bold"
+                       Foreground="{DynamicResource TextPrimary}"/>
+            <TextBlock x:Name="DialogHintText"
+                       FontSize="12"
+                       TextWrapping="Wrap"
+                       Foreground="{DynamicResource TextSecondary}"
+                       Margin="0,8,0,16"/>
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*"/>
+                    <ColumnDefinition Width="Auto"/>
+                </Grid.ColumnDefinitions>
+                <TextBox x:Name="PathTextBox"
+                         Style="{StaticResource InputTextBoxStyle}"
+                         MinHeight="36"
+                         VerticalContentAlignment="Center"
+                         ToolTip="本地文件夹或 UNC 网络路径"/>
+                <Button x:Name="BrowseButton"
+                        Grid.Column="1"
+                        MinWidth="88"
+                        Margin="10,0,0,0"
+                        Style="{StaticResource SecondaryButtonStyle}"
+                        Click="BrowseButton_Click"/>
+            </Grid>
+            <StackPanel Orientation="Horizontal"
+                        HorizontalAlignment="Right"
+                        Margin="0,22,0,0">
+                <Button x:Name="OkButton"
+                        MinWidth="88"
+                        Margin="0,0,10,0"
+                        IsDefault="True"
+                        Style="{StaticResource PrimaryButtonStyle}"
+                        Click="OkButton_Click"/>
+                <Button x:Name="CancelButton"
+                        MinWidth="88"
+                        IsCancel="True"
+                        Style="{StaticResource SecondaryButtonStyle}"/>
+            </StackPanel>
+        </StackPanel>
+    </Border>
+</Window>
